feat: Add delete functionality for notes in TopicPage
